modified np.float() -> np.float32
prettify

+++ detector/tracker/tracker/matching.py
@@ -1,12 +1,11 @@
-import cv2
+import numpy as np
 import numpy as np
 import scipy
-from scipy.spatial.distance import cdist
-from scipy.optimize import linear_sum_assignment
-
 from cython_bbox import bbox_overlaps as bbox_ious
+from scipy.optimize import linear_sum_assignment
+from scipy.spatial.distance import cdist
 from tracker.utils import kalman_filter
-import time
+
 
 def merge_matches(m1, m2, shape):
     O,P,Q = shape
@@ -61,13 +60,13 @@ def ious(atlbrs, btlbrs):
 
     :rtype ious np.ndarray
     """
-    ious = np.zeros((len(atlbrs), len(btlbrs)), dtype=np.float)
+    ious = np.zeros((len(atlbrs), len(btlbrs)), dtype=np.float32)
     if ious.size == 0:
         return ious
 
     ious = bbox_ious(
-        np.ascontiguousarray(atlbrs, dtype=np.float),
-        np.ascontiguousarray(btlbrs, dtype=np.float)
+        np.ascontiguousarray(atlbrs, dtype=np.float32),
+        np.ascontiguousarray(btlbrs, dtype=np.float32)
     )
 
     return ious
@@ -101,10 +100,10 @@ def embedding_distance(tracks, detections, metric='cosine'):
     :return: cost_matrix np.ndarray
     """
 
-    cost_matrix = np.zeros((len(tracks), len(detections)), dtype=np.float)
+    cost_matrix = np.zeros((len(tracks), len(detections)), dtype=np.float32)
     if cost_matrix.size == 0:
         return cost_matrix
-    det_features = np.asarray([track.curr_feat for track in detections], dtype=np.float)
+    det_features = np.asarray([track.curr_feat for track in detections], dtype=np.float32)
     for i, track in enumerate(tracks):
         cost_matrix[i, :] = np.maximum(0.0, cdist(track.smooth_feat.reshape(1,-1), det_features, metric))
     return cost_matrix
